refactor: apply single patch commit to new structure
